feat: Make log feature optional
- Add log feature flag
- Move log specific code to log module
- Add Level and LevelFilter associated types on LogLevel trait
- Duplicate Verbosity struct with cfg flags as there's no easy way to
  have an optional default for an associated type

BREAKING CHANGE: The log crate is now an optional dependency, enabled by
default. The `LogLevel` trait now has associated types for `Level` and
`LevelFilter` instead of these being the log crate types. The
`log::Level` and `log::LevelFilter` types are now re-exported from the
log module rather than the top-level module. This is a breaking change
that will affect users who have implemented the `LogLevel` trait or
imported the `Level` and `LevelFilter` types from the top-level module.
