+++ Cargo.toml
@@ -113,9 +113,13 @@ pre-release-replacements = [
 [badges]
 codecov = { repository = "clap-rs/clap-verbosity-flag" }
 
+[features]
+default = ["log"]
+log = ["dep:log"]
+
 [dependencies]
-log = "0.4.1"
 clap = { version = "4.0.0", default-features = false, features = ["std", "derive"] }
+log = { version = "0.4.1", optional = true }
 
 [dev-dependencies]
 clap = { version = "4.5.4", default-features = false, features = ["help", "usage"] }
